requirements.txt
torch
transformers==4.35.2
tweepy==4.14.0
pandas==2.1.3
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
datasets==2.15.0
accelerate==0.24.1
tensorboard==2.15.1
bitsandbytes==0.41.1
aiohttp==3.9.1
sentencepiece==0.1.99  # Añadido para LlamaTokenizer
protobuf==4.25.1       # Dependencia de sentencepiece